Updated Recipe hierarchy
