+++ src/main/java/me/shadowalzazel/mcodyssey/datagen/RecipeManager.kt
@@ -1,10 +1,9 @@
 package me.shadowalzazel.mcodyssey.datagen
 
-import me.shadowalzazel.mcodyssey.datagen.recipes.MerchantRecipes
-import me.shadowalzazel.mcodyssey.datagen.recipes.UtilityCrafting
-import me.shadowalzazel.mcodyssey.datagen.recipes.UtilitySmithing
+import me.shadowalzazel.mcodyssey.datagen.recipes.*
+import me.shadowalzazel.mcodyssey.datagen.recipes.crafting.EquipmentRecipes
 import me.shadowalzazel.mcodyssey.datagen.recipes.crafting.FoodRecipes
-import me.shadowalzazel.mcodyssey.datagen.recipes.creators.GeneratedRecipes
+import me.shadowalzazel.mcodyssey.datagen.recipes.crafting.ItemRecipes
 import me.shadowalzazel.mcodyssey.datagen.recipes.creators.ToolRecipeCreator
 import me.shadowalzazel.mcodyssey.datagen.recipes.creators.WeaponRecipeCreator
 import org.bukkit.inventory.Recipe
@@ -15,14 +14,16 @@ class RecipeManager {
         val recipeList: MutableList<Recipe> = mutableListOf()
         recipeList.apply {
             addAll(GeneratedRecipes().getRecipes())
+            addAll(ItemRecipes().getRecipes())
+            addAll(EquipmentRecipes().getRecipes())
             addAll(FoodRecipes().getRecipes())
+            addAll(CookingRecipes().getRecipes())
             addAll(MerchantRecipes().getRecipes())
-            addAll(UtilitySmithing().getRecipes())
-            addAll(ToolRecipeCreator().generateToolRecipes())
-            addAll(WeaponRecipeCreator().generateWeaponRecipes())
+            addAll(SmeltingRecipes().getRecipes())
             addAll(UtilityCrafting().getRecipes())
+            addAll(UtilitySmithing().getRecipes())
         }
         return recipeList
     }
-
+    // hierarchy
 }
