+++ src/main/java/me/shadowalzazel/mcodyssey/datagen/recipes/creators/RocketRecipeCreator.kt
@@ -11,6 +11,9 @@ import org.bukkit.inventory.recipe.CraftingBookCategory
 
 class RocketRecipeCreator {
 
+
+
+
     internal fun blazingRocketsRecipe(tier: Int, tierName: String): ShapelessRecipe {
         // New Rocket
         val rocket = Item.BLAZING_ROCKET.newItemStack(3)
